tensorflow/requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow-cpu
python-multipart